requirements.txt
appdirs==1.4.4
appnope==0.1.0
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.1.0
backcall==0.1.0
beautifulsoup4==4.10.0
BitCash==0.6.4
celery==5.1.2
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
daphne==3.0.1
decorator==4.4.1
distlib==0.3.1
Django==3.0.14
django-cors-headers==3.7.0
django-jinja==2.6.0
django-postgres-extra==2.0.2
django-rest-hooks==1.5.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-dynamic-raw-id==2.8
drf-yasg==1.17.1
filelock==3.0.12
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
humanize==0.5.1
idna==2.9
importlib-metadata==1.5.0
importlib-resources==3.0.0
inflection==0.5.1
ipython==7.12.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.16.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
plumbum==1.6.9
prometheus-client==0.8.0
prompt-toolkit==3.0.3
psycopg2==2.8.4
ptyprocess==0.6.0
py==1.9.0
Pillow==8.3.2
Pygments==2.5.2
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.0.1
pytest-cov==2.7.1
pytest-django==3.5.1
pytest-mock==1.10.4
python-decouple==3.3
pytz==2019.3
redis==3.4.1
requests==2.23.0
rpyc==4.1.5
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.0
sseclient==0.0.24
supervisor==4.1.0
tornado==6.0.4
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
web3==5.28.0
whitenoise==5.0.1
zipp==3.0.0
zope.event==4.4
zope.interface==5.1.0
grpcio==1.34.0
grpcio-tools==1.34.0
channels-redis==3.2.0
channels==3.0.2
cashaddress==1.0.6
